[ntuple] factor out rfield class test
