+++ tree/ntuple/v7/test/ntuple.cxx
@@ -220,21 +220,6 @@ TEST(RNTuple, TypeName) {
                 ROOT::Experimental::RField<CustomStruct>::TypeName().c_str());
 }
 
-namespace {
-class RNoDictionary {};
-} // namespace
-
-TEST(RNTuple, TClass) {
-   auto modelFail = RNTupleModel::Create();
-   EXPECT_THROW(modelFail->MakeField<RNoDictionary>("nodict"), std::runtime_error);
-
-   auto model = RNTupleModel::Create();
-   auto ptrKlass = model->MakeField<CustomStruct>("klass");
-
-   FileRaii fileGuard("test_ntuple_tclass.root");
-   auto ntuple = RNTupleWriter::Recreate(std::move(model), "f", fileGuard.GetPath());
-}
-
 TEST(RNTuple, Descriptor)
 {
    RNTupleDescriptorBuilder descBuilder;
