+++ tree/ntuple/v7/test/rfield_class.cxx
@@ -1,0 +1,16 @@
+#include "ntuple_test.hxx"
+
+namespace {
+class RNoDictionary {};
+} // namespace
+
+TEST(RNTuple, TClass) {
+   auto modelFail = RNTupleModel::Create();
+   EXPECT_THROW(modelFail->MakeField<RNoDictionary>("nodict"), std::runtime_error);
+
+   auto model = RNTupleModel::Create();
+   auto ptrKlass = model->MakeField<CustomStruct>("klass");
+
+   FileRaii fileGuard("test_ntuple_tclass.root");
+   auto ntuple = RNTupleWriter::Recreate(std::move(model), "f", fileGuard.GetPath());
+}
